refactor: split oversized rust modules
